mov threading to command
