+++ kernel/include/system_call.h
@@ -1,0 +1,7 @@
+#ifndef _SYSTEM_CALL_H
+#define _SYSTEM_CALL_H
+
+void sys_schedule();
+void sys_uart_read();
+
+#endif
