+++ kernel/src/exception/system_call.S
@@ -3,3 +3,7 @@
 .global sys_schedule
 sys_schedule:
     svc 0
+
+.global sys_schedule
+sys_uart_read:
+    svc 1
